etm: use lock-free MPSC node-based queue instead of immutable queue.
There were a few problems with the old immutable queue model that made
it not scale as well as it needed to. In cases where the writers are
faster than the readers the immutable queue would be tail heavy and
cause high memory usage. Switching to a linked list backed queue fixes
this problem by allowing processed nodes to be garbage collected
sooner.

This queue is also significantly faster than the old immutable queue
allowing for more messages to be processed in the same amount of time.